preprocessor/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "cryptography>=46.0.6",
    "fastapi>=0.135.2",
    "httpx>=0.28.1",
    "lxml>=6.0.2",
    "pandas>=3.0.1",
    "pydantic>=2.12.5",
    "pymupdf>=1.27.2.2",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "requests>=2.33.0",
    "uvicorn[standard]>=0.42.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
